Get tests running with modern dependencies
Most deps are not yet pinned to modern versions or updated (ZODB3),
but the two that we do require, zope.testbrowser and mimetypes, are
pinned.

The way that zope.testbrowser works is completely different than it
did before. So now we have two separate layers for now. But that
demonstrates a *much* nicer way we can do our tests in the future,
probably: inherit from zope.app.appsetup.testlayer.ZODBLayer

Add initial .travis.yml for 2.7; PyPy2 does not yet work.

Add .coveragerc and support for coveralls on Travis.